requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
python-telegram-bot
requests
plyer
pydantic
